Fazendo crud de Conta Bancária de Fornecedor

+++ src/DEKL.CP.Domain/Contracts/Entities/IBankAgency.cs
@@ -2,7 +2,7 @@
 {
     public interface IBankAgency
     {
-        int Number { get; set; }
+        short Number { get; set; }
         string ManagerName { get; set; }
         string PhoneNumber { get; set; }
         string Email { get; set; }
